Move orche yaml to core module
