+++ shardingsphere-control-panel/shardingsphere-orchestration/shardingsphere-orchestration-core/shardingsphere-orchestration-core-common/src/main/java/org/apache/shardingsphere/orchestration/core/common/yaml/config/YamlOrchestrationCenterConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.orchestration.repository.common.yaml.config;
+package org.apache.shardingsphere.orchestration.core.common.yaml.config;
 
 import lombok.Getter;
 import lombok.Setter;
