+++ shardingsphere-control-panel/shardingsphere-orchestration/shardingsphere-orchestration-core/shardingsphere-orchestration-core-config/src/test/resources/yaml/data-source.yaml
@@ -15,7 +15,8 @@
 # limitations under the License.
 #
 
-master_ds: !!org.apache.shardingsphere.orchestration.core.common.configuration.YamlDataSourceConfiguration
+master_ds: !!org.apache.shardingsphere.orchestration.core.common.yaml.config.YamlDataSourceConfiguration
+  
   dataSourceClassName: com.zaxxer.hikari.HikariDataSource
   props:
     url: jdbc:mysql://localhost:3306/demo_ds_master
