+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -77,7 +77,7 @@
     },
     {
       "name": "spring.shardingsphere.orchestration",
-      "type": "java.util.Map<java.lang.String,org.apache.shardingsphere.orchestration.repository.common.yaml.config.YamlOrchestrationCenterConfiguration>",
+      "type": "java.util.Map<java.lang.String,org.apache.shardingsphere.orchestration.core.common.yaml.config.YamlOrchestrationCenterConfiguration>",
       "sourceType": "org.apache.shardingsphere.spring.boot.orchestration.common.OrchestrationSpringBootRootConfiguration",
       "description": "Customize ShardingSphere orchestration instance."
     },
