+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-orchestration-spring/shardingsphere-jdbc-orchestration-spring-boot-starter/src/test/resources/yaml/data-source.yaml
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-dataSource: !!org.apache.shardingsphere.orchestration.core.common.configuration.YamlDataSourceConfiguration
+dataSource: !!org.apache.shardingsphere.orchestration.core.common.yaml.config.YamlDataSourceConfiguration
   dataSourceClassName: org.apache.commons.dbcp2.BasicDataSource
   props:
     url: jdbc:h2:mem:ds;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
